chore: restructure eval
