+++ eval/scripts/_paths.py
@@ -1,7 +1,7 @@
 import sys
 from pathlib import Path
 
-ERASING_ROOT = Path(__file__).resolve().parents[1]
+ERASING_ROOT = Path(__file__).resolve().parents[2]
 
 if str(ERASING_ROOT) not in sys.path:
     sys.path.insert(0, str(ERASING_ROOT))
