+++ eval/scripts/generate-images.py
