+++ eval/scripts/record_fid.py
@@ -8,7 +8,7 @@ import argparse
 from _paths import ERASING_ROOT  # noqa: F401
 
 from eval.layout import experiment_paths
-from eval.results import load_results, save_results, update_fid_scores
+from eval.results_io import load_results, save_results, update_fid_scores
 
 
 def main() -> None:
