+++ eval/scripts/sld-generate-images.py
